Update
Temperature got updated to f64 for better accuracy

+++ src/models/storage.rs
@@ -1,5 +1,5 @@
 pub struct Storage {
-	pub temperature: f32,
+	pub temperature: f64,
 	pub used: f64,
 	pub total: f64
 }
